# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: PC Builder Agent Product API
  version: 1.0.0
  license:
    name: MIT
    identifier: MIT
  description: |
    产品 API 契约。P7 会话、run、SSE 与配置读取、P9 分享，以及本地 Supabase Auth 账号接口均已实现。
    浏览器不得绕过本契约直接依赖 ADK/A2A 线格式。
servers:
  - url: /
    description: 同源入口;本地由 Next.js rewrite 转发至 Go :8082
...
paths:
...
  /api/v1/sessions/{session_id}:
    parameters:
      - $ref: "#/components/parameters/SessionID"
    patch:
      summary: 重命名或归档/恢复自己的对话，不触发模型运行
      tags: [sessions]
      operationId: updateSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              minProperties: 1
              properties:
                title: {type: string, minLength: 1, maxLength: 80}
                archived: {type: boolean}
      responses:
        "200":
          description: 更新后的会话
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Session"}
        "4XX":
          $ref: "#/components/responses/Problem"
    delete:
      summary: 永久删除自己的对话及消息、配置版本、分享；运行中返回 409
      tags: [sessions]
      operationId: deleteSession
      responses:
        "204":
          description: 已删除
        "4XX":
          $ref: "#/components/responses/Problem"
    get:
      summary: 读取一个产品会话
      tags: [sessions]
      operationId: getSession
      responses:
        "200":
          description: 会话详情、消息、待确认需求和活动 run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "4XX":
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    anonymousCookie:
      type: apiKey
      in: cookie
      name: pcb_anonymous_id
    authSessionCookie:
      type: apiKey
      in: cookie
      name: pcb_auth_session
  parameters:
...
    SessionID:
      name: session_id
      in: path
      required: true
      schema: {type: string, minLength: 1, maxLength: 128}
...
  responses:
    Problem:
      description: RFC problem details;404 同时用于不存在与无权访问
      content:
        application/problem+json:
          schema: {$ref: "#/components/schemas/Problem"}
  schemas:
...
    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status, code, request_id]
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string}
        status: {type: integer, minimum: 400, maximum: 599}
        detail: {type: string}
        instance: {type: string, format: uri-reference}
        code:
          type: string
          enum:
            - invalid_request
            - not_found
            - session_busy
            - invalid_session_phase
            - schema_validation_failed
            - upstream_unavailable
            - context_expired
            - run_timeout
            - run_interrupted
            - generation_failed
            - events_expired
            - internal_error
            - auth_disabled
            - auth_invalid_credentials
            - auth_email_exists
            - auth_weak_password
            - auth_session_expired
            - auth_unavailable
        request_id: {type: string}
...
    SessionPhase:
      type: string
      enum: [collecting, requirement_ready, building, ready, changing, error]
    SessionSummary:
      type: object
      additionalProperties: false
      required: [schema_version, id, title, phase, created_at, updated_at, version_count]
      properties:
        schema_version: {type: integer, const: 1}
        id: {type: string}
        title: {type: string}
        archived: {type: boolean, description: 是否已归档；旧客户端未提供时视为 false}
        phase: {$ref: "#/components/schemas/SessionPhase"}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        version_count: {type: integer, minimum: 0}
    Session:
      allOf:
        - $ref: "#/components/schemas/SessionSummary"
        - type: object
          required: [messages, pending_requirement, requirement_state, requirement_status, confirmed_requirement_state, confirmed_requirement, confirmed_at, missing_fields, active_run, last_error, recovery_phase, degraded]
          properties:
            messages:
              type: array
              items: {$ref: "#/components/schemas/Message"}
            pending_requirement:
              oneOf:
                - $ref: "#/components/schemas/RequirementSpec"
                - type: "null"
            requirement_state:
              oneOf:
                - $ref: "#/components/schemas/RequirementState"
                - type: "null"
            requirement_status:
              type: string
              enum: [collecting, ready_to_confirm, confirmed, modified]
              description: 根据当前有效需求与已确认快照计算；仅讨论备选或无有效修改不会被标为 modified。
            confirmed_requirement_state:
              oneOf:
                - $ref: "#/components/schemas/RequirementState"
                - type: "null"
            confirmed_requirement:
              oneOf:
                - $ref: "#/components/schemas/RequirementSpec"
                - type: "null"
            confirmed_at:
              oneOf:
                - {type: string, format: date-time}
                - {type: "null"}
            missing_fields:
              type: array
              items: {type: string}
              description: 当前需求生成前仍需补充或消解冲突的必要字段。
            active_run:
              oneOf:
                - $ref: "#/components/schemas/Run"
                - type: "null"
            last_error:
              oneOf:
                - $ref: "#/components/schemas/Problem"
                - type: "null"
            recovery_phase:
              oneOf:
                - type: string
                  enum: [collecting, requirement_ready, ready]
                - type: "null"
            degraded: {type: boolean}
    Message:
      type: object
      additionalProperties: false
      required: [schema_version, id, role, content, created_at]
      properties:
        schema_version: {type: integer, const: 1}
        id: {type: string, format: uuid}
        role: {type: string, enum: [user, assistant]}
        content: {type: string}
        display_content:
          type: string
          description: 服务端生成的用户可读摘要；展示与复制优先使用非空摘要，原始 content 保留供追溯。
        run_id:
          oneOf:
            - {type: string, format: uuid}
            - {type: "null"}
        created_at: {type: string, format: date-time}
...
    Run:
      type: object
      additionalProperties: false
      required: [schema_version, id, session_id, kind, status, started_at, events_url]
      properties:
        schema_version: {type: integer, const: 1}
        id: {type: string, format: uuid}
        session_id: {type: string}
        kind: {type: string, enum: [screening, build, change]}
        status: {type: string, enum: [running, succeeded, failed, interrupted]}
        started_at: {type: string, format: date-time}
        finished_at:
          oneOf:
            - {type: string, format: date-time}
            - {type: "null"}
        error:
          oneOf:
            - $ref: "#/components/schemas/Problem"
            - type: "null"
        events_url: {type: string, format: uri-reference}
    RequirementSource:
      type: object
      additionalProperties: false
      required: [kind, message_id, quote]
      properties:
        kind: {type: string, enum: [chat, edit]}
        message_id: {type: string}
        quote: {type: string}
    RequirementField:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        value: {}
        status: {type: string, enum: [unknown, active, removed, conflict]}
        strength: {type: string, enum: [must, prefer]}
        scope: {type: string, enum: [session, temporary]}
        source: {$ref: "#/components/schemas/RequirementSource"}
        previous: {$ref: "#/components/schemas/RequirementField"}
...
    RequirementAlternative:
      type: object
      additionalProperties: false
      required: [field, value, strength, scope, source]
      properties:
        field: {type: string}
        value: {}
        strength: {type: string, enum: [must, prefer]}
        scope: {type: string, enum: [session, temporary]}
        source: {$ref: "#/components/schemas/RequirementSource"}
    RequirementChange:
      type: object
      additionalProperties: false
      required: [revision, op, field, source]
      properties:
        revision: {type: integer, minimum: 1}
        op: {type: string, enum: [set, remove, restore, alternative, conflict]}
        field: {type: string}
        before: {$ref: "#/components/schemas/RequirementField"}
        after: {$ref: "#/components/schemas/RequirementField"}
        source: {$ref: "#/components/schemas/RequirementSource"}
    RequirementState:
      type: object
      additionalProperties: false
      required: [schema_version, revision, fields, alternatives, changes, history]
      description: 仅当前装机会话的需求真值；未知、撤销、冲突字段不被程序默认值覆盖；临时例外可恢复之前的字段。备选讨论不改变当前有效需求。
      properties:
        schema_version: {type: integer, const: 1}
        revision: {type: integer, minimum: 0}
        fields:
          type: object
          additionalProperties: {$ref: "#/components/schemas/RequirementField"}
        alternatives:
          type: array
          items: {$ref: "#/components/schemas/RequirementAlternative"}
        changes:
          type: array
          items: {$ref: "#/components/schemas/RequirementChange"}
        history:
          type: array
          items: {$ref: "#/components/schemas/RequirementChange"}
    RequirementSpec:
      type: object
      additionalProperties: false
      required: [schema_version, budget_cny, use_case]
      properties:
        schema_version: {type: integer, const: 1}
        budget_cny: {type: integer, minimum: 1}
        budget_flex: {type: number, minimum: 0, maximum: 0.3, default: 0.1}
        use_case:
          type: object
          additionalProperties: false
          required: [type]
          properties:
            type: {type: string, enum: [gaming, productivity, general]}
            titles: {type: array, items: {type: string}, default: []}
            resolution: {type: string, enum: [1080p, 2K, 4K]}
            fps_target: {type: integer, minimum: 1}
          allOf:
            - if:
                properties:
                  type: {const: gaming}
                required: [type]
              then:
                properties:
                  resolution: {type: string, enum: [1080p, 2K, 4K]}
                required: [resolution]
        size_pref: {type: string, enum: [atx, matx, itx, any], default: any}
        noise_pref: {type: string, enum: [silent, normal, any], default: any}
        brand_pref:
          type: object
          additionalProperties: false
          properties:
            cpu: {type: string, enum: [any, intel, amd], default: any}
            gpu: {type: string, enum: [any, nvidia, amd], default: any}
        existing_parts:
          type: array
          items: {$ref: "#/components/schemas/PartCategory"}
          default: []
        budget_basis:
          type: string
          enum: [new_purchase, full_build]
          description: 有已有件时必须明确；新增购买费用或整机参考总价。
        owned_parts:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [category, model]
            properties:
              category: {$ref: "#/components/schemas/PartCategory"}
              model: {type: string, minLength: 1}
              quantity: {type: integer, minimum: 1, maximum: 8, default: 1}
          description: 每品类一个准确型号；仅 SSD 可多件，其余数量为 1。程序执行前精确绑定并锁定。
        priority:
          type: array
          items: {$ref: "#/components/schemas/PartCategory"}
          default: []
        notes: {type: string, default: ""}
        constraint_strengths:
          type: object
          additionalProperties: {type: string, enum: [must, prefer]}
          description: 当前有效字段的明确强度，键采用 brand_pref.cpu、size_pref 等需求字段路径；缺失时兼容旧需求语义，不据此认定为用户已表达。
        requirement_details:
          type: object
          additionalProperties: false
          properties:
            appearance: {type: string}
            recipient: {type: string}
          description: 当前会话的有效外观与装机对象说明；不含已撤销信息、备选方案或长期画像。
    PartCategory:
      type: string
      enum: [cpu, gpu, motherboard, memory, ssd, psu, case, cooler]
...
security:
  - anonymousCookie: []
  - authSessionCookie: []